refactor(dogstatsd): renamed replay file
